requirements.txt
tqdm
nltk
noisereduce
scipy
sounddevice
brainflow
absl-py
matplotlib
soundfile
librosa
numpy
pyworld
openai-whisper
absl-py
librosa
soundfile
scipy 
scikit-learn 
numba 
jiwer 
unidecode
torch